+++ Kms Cloud Database/UserDataTotalDistance.cs
@@ -18,9 +18,9 @@ namespace Kms.Cloud.Database
         public System.DateTime Timestamp { get; set; }
         public long TotalSteps { get; set; }
         public long TotalDistance { get; set; }
-        public long TotalKcal { get; set; }
-        public long TotalCo2 { get; set; }
-        public long TotalCash { get; set; }
+        public double TotalKcal { get; set; }
+        public double TotalCo2 { get; set; }
+        public double TotalCash { get; set; }
     
         public virtual User User { get; set; }
     }
